move logic back to endpoints
