+++ apps/dotcom/sync-worker/src/routes/tla/unpublishFile.ts
@@ -1,6 +1,7 @@
 import { IRequest } from 'itty-router'
+import { getR2KeyForRoom } from '../../r2'
 import { Environment } from '../../types'
-import { getTldrawAppDurableObject } from '../../utils/tla/getTldrawAppDurableObject'
+import { getTldrawAppFileRecord } from '../../utils/tla/getTldrawAppFileRecord'
 import { getUserIdFromRequest } from '../../utils/tla/permissions'
 
 // Unpublish a file.
@@ -15,10 +16,26 @@ export async function unpublishFile(request: IRequest, env: Environment): Promis
 		return Response.json({ error: true, message: 'No user' }, { status: 401 })
 	}
 
-	const app = getTldrawAppDurableObject(env)
-
 	try {
-		await app.unpublishFile(roomId, userId)
+		const file = await getTldrawAppFileRecord(roomId, env)
+
+		if (!file) {
+			throw Error('not-found')
+		}
+
+		// A user can only publish their own files
+		if (file.ownerId !== userId) {
+			throw Error('forbidden')
+		}
+
+		// Create a new slug for the published room
+		await env.SNAPSHOT_SLUG_TO_PARENT_SLUG.delete(file.publishedSlug)
+
+		// Bang the snapshot into the database
+		await env.ROOM_SNAPSHOTS.delete(
+			getR2KeyForRoom({ slug: `${roomId}/${file.publishedSlug}`, isApp: true })
+		)
+
 		return new Response(JSON.stringify({ error: false }))
 	} catch (e: any) {
 		return new Response(JSON.stringify({ error: true, message: e.message }), { status: 500 })
